+++ src/main/java/kitchenpos/order/dto/OrderTableResponse.java
@@ -1,9 +1,9 @@
-package kitchenpos.table.dto;
+package kitchenpos.order.dto;
 
 import java.util.Objects;
 
-import kitchenpos.table.domain.OrderTable;
-import kitchenpos.table.domain.TableGroup;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.TableGroup;
 
 public class OrderTableResponse {
