+++ src/main/java/kitchenpos/order/dto/TableGroupResponse.java
@@ -1,10 +1,10 @@
-package kitchenpos.table.dto;
+package kitchenpos.order.dto;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import kitchenpos.table.domain.TableGroup;
+import kitchenpos.order.domain.TableGroup;
 
 public class TableGroupResponse {
