+++ src/main/java/kitchenpos/order/repository/OrderTableRepository.java
@@ -1,11 +1,11 @@
-package kitchenpos.table.repository;
+package kitchenpos.order.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import kitchenpos.table.domain.OrderTable;
+import kitchenpos.order.domain.OrderTable;
 
 @Repository
 public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
