+++ src/main/java/kitchenpos/order/repository/TableGroupRepository.java
@@ -1,9 +1,9 @@
-package kitchenpos.table.repository;
+package kitchenpos.order.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import kitchenpos.table.domain.TableGroup;
+import kitchenpos.order.domain.TableGroup;
 
 @Repository
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
