+++ src/test/java/kitchenpos/order/domain/OrderTableTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.domain;
+package kitchenpos.order.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -10,8 +10,6 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import kitchenpos.order.domain.NumberOfGuests;
 
 class OrderTableTest {
